Presentation XML, concepts: https://github.com/metanorma/isodoc/issues/660
