+++ lib/isodoc/iso/presentation_xml_convert.rb
@@ -2,6 +2,7 @@ require_relative "init"
 require "isodoc"
 require_relative "presentation_xref"
 require_relative "presentation_bibdata"
+require_relative "presentation_section"
 require_relative "presentation_terms"
 require_relative "../../relaton/render/general"
 
@@ -20,39 +21,6 @@ module IsoDoc
           @suppressheadingnumbers = true
         end
         super
-      end
-
-      def section(docxml)
-        super
-        warning_for_missing_metadata(docxml)
-      end
-
-      def warning_for_missing_metadata(docxml)
-        @meta.get[:unpublished] or return
-        ret = warning_for_missing_metadata_create(docxml)
-        ret.empty? and return
-        warning_for_missing_metadata_post(docxml, ret)
-      end
-
-      def warning_for_missing_metadata_create(docxml)
-        ret = ""
-        docxml.at(ns("//bibdata/ext//secretariat")) or
-          ret += "<p>Secretariat is missing.</p>"
-        docxml.at(ns("//bibdata/ext//editorialgroup")) or
-          ret += "<p>Editorial groups are missing.</p>"
-        docxml.at(ns("//bibdata/date[@type = 'published' or @type = 'issued' " \
-                     "or @type = 'created']")) ||
-          docxml.at(ns("//bibdata/version/revision-date")) or
-          ret += "<p>Document date is missing.</p>"
-        ret
-      end
-
-      def warning_for_missing_metadata_post(docxml, ret)
-        id = UUIDTools::UUID.random_create
-        ret = "<review date='#{Date.today}' reviewer='Metanorma' id='_#{id}'>" \
-              "<p><strong>Metadata warnings:<strong></p> #{ret}</review>"
-        ins = docxml.at(ns("//sections//fmt-title")) or return
-        ins.add_first_child ret
       end
 
       def block(docxml)
@@ -79,26 +47,6 @@ module IsoDoc
         note docxml
       end
 
-      # Redo Amendment annex titles as numbered
-      def annex(isoxml)
-        amd?(isoxml) and @suppressheadingnumbers = @oldsuppressheadingnumbers
-        super
-        amd?(isoxml) and @suppressheadingnumbers = true
-      end
-
-      # Redo Amendment annex subclause titles as numbered
-      def clause(docxml)
-        super
-        docxml.xpath(ns("//annex//appendix")).each { |f| clause1(f) }
-        amd?(docxml) or return
-        @suppressheadingnumbers = @oldsuppressheadingnumbers
-        docxml.xpath(ns("//annex//clause | //annex//appendix")).each do |f|
-          f.xpath(ns("./fmt-title | ./fmt-xref-label")).each(&:remove)
-          clause1(f)
-        end
-        @suppressheadingnumbers = true
-      end
-
       def subfigure_delim
         "<span class='fmt-label-delim'>)</span>"
       end
@@ -120,25 +68,6 @@ module IsoDoc
         name.nil? and return
         div.span class: "example_label" do |_p|
           name.children.each { |n| parse(n, div) }
-        end
-      end
-
-      def clause1(node)
-        !node.at(ns("./title")) &&
-          !%w(sections preface bibliography).include?(node.parent.name) and
-          node["inline-header"] = "true"
-        super
-        clause1_section_prefix(node)
-      end
-
-      def clause1_section_prefix(node)
-        if node["type"] == "section" &&
-            c = node.at(ns("./fmt-title//span[@class = 'fmt-caption-delim']"))
-          c.add_first_child(":")
-          t = node.at(ns("./fmt-title"))
-          # French l10n needs tab to be treated as space
-          t.replace @i18n.l10n(to_xml(t).gsub("<tab/>", "<tab> </tab>"))
-            .gsub(%r{<tab>[^<]+</tab>}, "<tab/>")
         end
       end
 
@@ -231,56 +160,6 @@ module IsoDoc
         super
       end
 
-      def toc_title(docxml)
-        %w(amendment technical-corrigendum).include?(@doctype) and return
-        super
-      end
-
-      def middle_title(docxml)
-        @meta.get[:doctitlemain].nil? || @meta.get[:doctitlemain].empty? and
-          return
-        s = docxml.at(ns("//sections")) or return
-        ret = "#{middle_title_main}#{middle_title_amd}"
-        s.add_first_child ret
-      end
-
-      def middle_title_main
-        ret = "<span class='boldtitle'>#{@meta.get[:doctitleintro]}"
-        @meta.get[:doctitleintro] && @meta.get[:doctitlemain] and
-          ret += " &#x2014; "
-        ret += @meta.get[:doctitlemain]
-        @meta.get[:doctitlemain] && @meta.get[:doctitlepart] and
-          ret += " &#x2014; "
-        ret += "</span>#{middle_title_part}"
-        "<p class='zzSTDTitle1'>#{ret}</p>"
-      end
-
-      def middle_title_part
-        ret = ""
-        if a = @meta.get[:doctitlepart]
-          b = @meta.get[:doctitlepartlabel] and
-            ret += "<span class='nonboldtitle'>#{b}:</span> "
-          ret += "<span class='boldtitle'>#{a}</span>"
-        end
-        ret
-      end
-
-      def middle_title_amd
-        ret = ""
-        if a = @meta.get[:doctitleamdlabel]
-          ret += "<p class='zzSTDTitle2'>#{a}"
-          a = @meta.get[:doctitleamd] and ret += ": #{a}"
-          ret += "</p>"
-        end
-        a = @meta.get[:doctitlecorrlabel] and
-          ret += "<p class='zzSTDTitle2'>#{a}</p>"
-        ret
-      end
-
-      def move_norm_ref_to_sections(docxml)
-        amd?(docxml) or super
-      end
-
       def twitter_cldr_localiser_symbols
         { group: "&#xA0;", fraction_group: "&#xA0;",
           fraction_group_digits: 3 }
@@ -303,10 +182,6 @@ module IsoDoc
         else
           ""
         end
-      end
-
-      def enable_indexsect
-        true
       end
 
       def fn_ref_label(fnote)
