+++ spec/spec_helper.rb
@@ -53,6 +53,8 @@ end
 
 def strip_guid(xml)
   xml.gsub(%r{ id=['"](fn:|ftn)?_[^"']+['"]}, ' id="_"')
+    .gsub(%r{ from="_[^"]+"}, ' from="_"')
+    .gsub(%r{ to="_[^"]+"}, ' to="_"')
     .gsub(%r{ name="_[^"]+"}, ' name="_"')
     .gsub(%r{ original-id="_[^"]+"}, ' original-id="_"')
     .gsub(%r{ href="#(fn:|ftn)?_[^"]+"}, ' href="#_"')
